requirements.txt
Django==1.9.6
django-allauth==0.24.1
django-rest-auth==0.7.0
djangorestframework==3.3.2
MySQL-python==1.2.5
numpy==1.10.4
oauthlib==1.0.3
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
